+++ test/ui_tests.dart
@@ -18,16 +18,22 @@ import 'package:angular/angular.dart';
 import 'package:angular/mock/module.dart';
 
 //import 'package:angular_ui/position.dart';
+import 'package:angular_ui/transition.dart';
 import 'package:angular_ui/buttons.dart';
+import 'package:angular_ui/collapse.dart';
 
 //part 'tests/position_tests.dart';
+part 'tests/transition_tests.dart';
 part 'tests/buttons_tests.dart';
+part 'tests/collapse_tests.dart';
 
 void main() {
   print('Running unit tests for Angular UI library.');
   useHtmlEnhancedConfiguration();
   group('All Tests:', () {
 //    test('Position', () => positionTests());
+    test('Transition', () => transitionTests());
     test('Buttons', () => buttonsTests());
+//    test('Collapse', () => collapseTests());
   });
 }